wechat_agent_mcp/pyproject.toml
[project]
name = "wechat-agent-mcp"
version = "0.1.0"
description = "A WeChat Official Account Agent based on MCP and Skills"
authors = [
    { name = "Antigravity", email = "assistant@google.com" }
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "mcp>=0.1.0",
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "rich>=13.0.0",
    "Pillow>=10.0.0",
    "tenacity>=8.0.0",
]
readme = "README.md"
requires-python = ">=3.10"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "skills*", "models*", "lib*", "utils*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
